+++ service/controller/resource/tccpn/create_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/giantswarm/aws-operator/service/internal/hamaster"
 	"github.com/giantswarm/aws-operator/service/internal/images"
 	"github.com/giantswarm/aws-operator/service/internal/recorder"
+	"github.com/giantswarm/aws-operator/service/internal/releases"
 	"github.com/giantswarm/aws-operator/service/internal/unittest"
 )
 
@@ -78,11 +79,24 @@ func Test_Controller_Resource_TCCPN_Template_Render(t *testing.T) {
 				}
 			}
 
+			var rel releases.Interface
+			{
+				c := releases.Config{
+					K8sClient: k,
+				}
+
+				rel, err = releases.New(c)
+				if err != nil {
+					t.Fatal(err)
+				}
+			}
+
 			var d *changedetection.TCCPN
 			{
 				c := changedetection.TCCPNConfig{
 					HAMaster: h,
 					Logger:   microloggertest.New(),
+					Releases: rel,
 				}
 
 				d, err = changedetection.NewTCCPN(c)