+++ service/controller/resource/tcnp/template/template_main_outputs.go
@@ -10,5 +10,7 @@ const TemplateMainOutputs = `
     Value: {{ .Outputs.Instance.Type }}
   OperatorVersion:
     Value: {{ .Outputs.OperatorVersion }}
+  ReleaseVersion:
+    Value: {{ .Outputs.ReleaseVersion }}
 {{- end -}}
 `